+++ pyutils/simplelogger.py
@@ -1,6 +1,5 @@
 import os
-import sys
-
+import logging
 import pyutils.shorthand as shd
 
 ######################
@@ -9,15 +8,10 @@ import pyutils.shorthand as shd
 ######################
 ######################
 
-LOG_LEVEL = None
-LOG_LEVEL_NORMAL = 0
-LOG_LEVEL_INFO = 1
-LOG_LEVEL_WARNING = 2
-LOG_LEVEL_ERROR = 3
-LOG_LEVEL_SUCCESS = 4
-LOG_LEVEL_NONE = 99
-LOG_INDENT = 0
 ErrorRaiseExcpetion = False
+
+logging.basicConfig(level=logging.INFO, datefmt='%y-%m-%d %H:%M:%S', format='%(message)s')
+logger = logging.getLogger()
 
 
 class SimpleLogger(object):
@@ -25,53 +19,36 @@ class SimpleLogger(object):
     def _color_message(message, color_code, bold=False):
         if shd.is_win():
             os.system('')
-        result = '\033[{}m{}\033[0m'.format(color_code, message)
-        if bold:
-            result = '\033[1m' + result
-        return result
+        bold_code = '\033[1m' if bold else ''
+        return f'{bold_code}\033[{color_code}m{message}\033[0m'
 
     @staticmethod
-    def _log(message, level=LOG_LEVEL_NORMAL, noident=False, bold=False):
-        global LOG_INDENT, LOG_LEVEL
+    def _preprocess_message(message):
+        if not shd.is_win():
+            message = message.replace('=>', '➜').replace('<=', '✔')
 
-        if LOG_LEVEL is None:
-            LOG_LEVEL = int(os.getenv('LoggerLevel', -1))
-
-        if level >= LOG_LEVEL:
-            if level == LOG_LEVEL_INFO:
-                message = SimpleLogger._color_message(message, 0, bold)
-
-            if level == LOG_LEVEL_WARNING:
-                message = SimpleLogger._color_message('warning: {}'.format(message), 33, bold)
-
-            if level == LOG_LEVEL_ERROR:
-                message = SimpleLogger._color_message('error: {}'.format(message), 31, bold)
-
-            if level == LOG_LEVEL_SUCCESS:
-                message = SimpleLogger._color_message('success: {}'.format(message), 32, bold)
-
-            if not shd.is_win():
-                message = message.replace('=>', '➜').replace('<=', '✔')
-
-            message += '\n'
-
-            pipe = sys.stdout if level != LOG_LEVEL_ERROR else sys.stderr
-
-            pipe.write(('' if noident else ('\t' * LOG_INDENT)) + message)
+        if message.endswith('\r\n') or message.endswith('\n'):
+            message = message.rstrip()
+        return message
 
     @staticmethod
-    def info(message, bold=False):
-        SimpleLogger._log(message, LOG_LEVEL_INFO, False, bold)
+    def info(message, _):
+        message = SimpleLogger._preprocess_message(message)
+        logger.info(message)
 
     @staticmethod
     def warning(message, bold=False):
-        SimpleLogger._log(message, LOG_LEVEL_WARNING, False, bold)
+        message = SimpleLogger._preprocess_message(message)
+        message = SimpleLogger._color_message(message, 33, bold)
+        logger.warning(message)
 
     @staticmethod
     def error(message, bold=False):
+        message = SimpleLogger._preprocess_message(message)
+        message = SimpleLogger._color_message(message, 31, bold)
         if ErrorRaiseExcpetion:
             raise Exception(message)
-        SimpleLogger._log(message, LOG_LEVEL_ERROR, False, bold)
+        logger.error(message)
 
 
 def info(message, bold=False):